infra/requirements.txt
# Track 1: Fine-tuning dependencies (install on GCP GPU instance)
# Unsloth handles torch, transformers, peft, bitsandbytes, trl internally
unsloth[cu124-torch250]
datasets>=2.18
pyyaml>=6.0
safetensors>=0.4
